Keep mixed optimizer transactions in shared primitives
